+++ test/smokes/metrics_fileinfo/expectations.rb
@@ -14,7 +14,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 7,
-          last_committed_at: "2021-01-01T10:00:00+09:00"
+          last_committed_at: "2021-01-01T10:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       }
@@ -34,7 +38,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: nil,
-          last_committed_at: "2021-01-01T13:00:00+09:00"
+          last_committed_at: "2021-01-01T13:00:00+09:00",
+          number_of_commits: 1,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       }, {
@@ -45,7 +53,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: nil,
-          last_committed_at: "2021-01-01T14:00:00+09:00"
+          last_committed_at: "2021-01-01T14:00:00+09:00",
+          number_of_commits: 1,
+          occurrence: 1,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       }
@@ -65,7 +77,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 2,
-          last_committed_at: "2021-01-01T11:00:00+09:00"
+          last_committed_at: "2021-01-01T11:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       }
@@ -85,7 +101,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 1,
-          last_committed_at: "2021-01-01T12:00:00+09:00"
+          last_committed_at: "2021-01-01T12:00:00+09:00",
+          number_of_commits: 3,
+          occurrence: 1,
+          additions: 1,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -97,7 +117,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 1,
-          last_committed_at: "2021-01-01T13:00:00+09:00"
+          last_committed_at: "2021-01-01T13:00:00+09:00",
+          number_of_commits: 3,
+          occurrence: 1,
+          additions: 1,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -109,7 +133,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 1,
-          last_committed_at: "2021-01-01T11:00:00+09:00"
+          last_committed_at: "2021-01-01T11:00:00+09:00",
+          number_of_commits: 3,
+          occurrence: 1,
+          additions: 1,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -121,7 +149,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 7,
-          last_committed_at: "2021-01-01T10:00:00+09:00"
+          last_committed_at: "2021-01-01T10:00:00+09:00",
+          number_of_commits: 3,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       }
@@ -141,7 +173,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 9,
-          last_committed_at: "2021-01-01T12:00:00+09:00"
+          last_committed_at: "2021-01-01T12:00:00+09:00",
+          number_of_commits: 2,
+          occurrence: 2,
+          additions: 2,
+          deletions: 0
         },
         git_blame_info: nil
       }
@@ -161,7 +197,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 2,
-          last_committed_at: "2021-01-01T14:00:00+09:00"
+          last_committed_at: "2021-01-01T14:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -173,7 +213,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 1,
-          last_committed_at: "2021-01-01T14:00:00+09:00"
+          last_committed_at: "2021-01-01T14:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -185,7 +229,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 5,
-          last_committed_at: "2021-01-01T14:00:00+09:00"
+          last_committed_at: "2021-01-01T14:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -197,7 +245,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 5,
-          last_committed_at: "2021-01-01T14:00:00+09:00"
+          last_committed_at: "2021-01-01T14:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -209,7 +261,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 5,
-          last_committed_at: "2021-01-01T14:00:00+09:00"
+          last_committed_at: "2021-01-01T14:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -221,7 +277,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 6,
-          last_committed_at: "2021-01-01T14:00:00+09:00"
+          last_committed_at: "2021-01-01T14:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       }
@@ -241,7 +301,11 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 7,
-          last_committed_at: "2021-01-01T10:00:00+09:00"
+          last_committed_at: "2021-01-01T10:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
         },
         git_blame_info: nil
       },
@@ -253,7 +317,123 @@ s.add_test_with_git_metadata(
         links: [],
         object: {
           lines_of_code: 2,
-          last_committed_at: "2021-01-01T10:00:00+09:00"
+          last_committed_at: "2021-01-01T10:00:00+09:00",
+          number_of_commits: 0,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
+        },
+        git_blame_info: nil
+      }
+    ],
+    analyzer: { name: "Metrics File Info", version: default_version } }
+)
+
+s.add_test_with_git_metadata(
+  "churn",
+  { type: "success",
+    issues: [
+      {
+        id: "metrics_fileinfo",
+        path: "foo/test.txt",
+        location: nil,
+        message: "foo/test.txt: loc = 5, last commit datetime = 2021-03-04T04:23:45+00:00",
+        links: [],
+        object: {
+          lines_of_code: 5,
+          last_committed_at: "2021-03-04T04:23:45+00:00",
+          number_of_commits: 4,
+          occurrence: 2,
+          additions: 5,
+          deletions: 0
+        },
+        git_blame_info: nil
+      },
+      {
+        id: "metrics_fileinfo",
+        path: "image.png",
+        location: nil,
+        message: "image.png: loc = (no info), last commit datetime = 2021-03-04T04:23:45+00:00",
+        links: [],
+        object: {
+          lines_of_code: nil,
+          last_committed_at: "2021-03-04T04:23:45+00:00",
+          number_of_commits: 4,
+          occurrence: 1,
+          additions: 0,
+          deletions: 0
+        },
+        git_blame_info: nil
+      },
+      {
+        id: "metrics_fileinfo",
+        path: "sample.c",
+        location: nil,
+        message: "sample.c: loc = 2, last commit datetime = 2021-03-05T05:23:45+00:00",
+        links: [],
+        object: {
+          lines_of_code: 2,
+          last_committed_at: "2021-03-05T05:23:45+00:00",
+          number_of_commits: 4,
+          occurrence: 3,
+          additions: 5,
+          deletions: 3
+        },
+        git_blame_info: nil
+      }
+    ],
+    analyzer: { name: "Metrics File Info", version: default_version } }
+)
+
+s.add_test_with_git_metadata(
+  "churn_days",
+  { type: "success",
+    issues: [
+      {
+        id: "metrics_fileinfo",
+        path: "bar/hello.java",
+        location: nil,
+        message: "bar/hello.java: loc = 2, last commit datetime = 2020-05-22T12:00:00+09:00",
+        links: [],
+        object: {
+          lines_of_code: 2,
+          last_committed_at: "2020-05-22T12:00:00+09:00",
+          number_of_commits: 112,
+          occurrence: 0,
+          additions: 0,
+          deletions: 0
+        },
+        git_blame_info: nil
+      },
+      {
+        id: "metrics_fileinfo",
+        path: "foo/test.txt",
+        location: nil,
+        message: "foo/test.txt: loc = 3, last commit datetime = 2020-06-20T12:00:00+09:00",
+        links: [],
+        object: {
+          lines_of_code: 3,
+          last_committed_at: "2020-06-20T12:00:00+09:00",
+          number_of_commits: 112,
+          occurrence: 111,
+          additions: 113,
+          deletions: 113
+        },
+        git_blame_info: nil
+      },
+      {
+        id: "metrics_fileinfo",
+        path: "sample.c",
+        location: nil,
+        message: "sample.c: loc = 1, last commit datetime = 2020-08-23T12:34:56+09:00",
+        links: [],
+        object: {
+          lines_of_code: 1,
+          last_committed_at: "2020-08-23T12:34:56+09:00",
+          number_of_commits: 112,
+          occurrence: 2,
+          additions: 1,
+          deletions: 3
         },
         git_blame_info: nil
       }