+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Misc:
 - `Runners::Config` check warnings [#2221](https://github.com/sider/runners/pull/2221)
 - **remark-lint** Use `.git` directory and add `remark-validate-links` to default ruleset [#2224](https://github.com/sider/runners/pull/2224)
 - Introduce common option `linter.<id>.dependencies` [#2211](https://github.com/sider/runners/pull/2211)
+- **Metrics File Info** Report the number of commits, additions, deletions used to measure code churn [#2184](https://github.com/sider/runners/pull/2184)
 
 ## 0.45.0